resulting image
The raw xlsx-dataframes have been adjusted and plotting with python. 
The following research question has been addressed: People from Michigan seem to like sports events, but how did the enthusiasm of the people develop over time to visit such sports events and how great is this enthusiasm compared to the rest of the world? In order to address this question a closer look on the Michigan stadium's hosting capacity has been done.